Treat missing conditionals as False
In #125 I added support for missing conditionals, but inside an And I made the assumption they defaulted to True. That's not actually what happens, and after experimenting, it seems that missing booleans (at least in the places I checked, namely `if`, `if/else`, `and`, `or` and `not`) simply default to `False`.